feat(qmk): move tap dance to userspace
