+++ users/haunt98/haunt98.c
@@ -1,0 +1,6 @@
+// Copyright 2023 haunt98 <hauvipapro@gmail.com>
+// SPDX-License-Identifier: MIT
+
+#include "haunt98.h"
+
+// Placebo for the future me
