+++ users/haunt98/haunt98.h
@@ -1,0 +1,9 @@
+// Copyright 2023 haunt98 <hauvipapro@gmail.com>
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+// https://github.com/qmk/qmk_firmware/blob/master/docs/feature_userspace.md
+
+#include QMK_KEYBOARD_H
+#include "tap_dance.h"
