+++ users/haunt98/rules.mk
@@ -22,3 +22,7 @@ TAP_DANCE_ENABLE = yes
 DEBOUNCE_TYPE = sym_eager_pk
 
 NKRO_ENABLE = yes
+
+# Custom C
+SRC += haunt98.c
+SRC += tap_dance.c
