+++ users/haunt98/tap_dance.h
@@ -1,0 +1,11 @@
+// Copyright 2023 haunt98 <hauvipapro@gmail.com>
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include QMK_KEYBOARD_H
+
+// https://github.com/qmk/qmk_firmware/blob/master/docs/feature_tap_dance.md
+enum custom_tap_dance {
+    TD_SAFE_BOOT,
+};
